+++ src/RestApi/BlogpostController.php
@@ -5,9 +5,11 @@ namespace Foodsharing\RestApi;
 use Foodsharing\Lib\Session;
 use Foodsharing\Modules\Blog\BlogGateway;
 use Foodsharing\Permissions\BlogPermissions;
+use Foodsharing\Utility\Sanitizer;
 use FOS\RestBundle\Controller\AbstractFOSRestController;
 use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Request\ParamFetcher;
+use Nelmio\ApiDocBundle\Annotation\Model;
 use OpenApi\Annotations as OA;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
@@ -16,18 +18,34 @@ use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 
 class BlogpostController extends AbstractFOSRestController
 {
-    private BlogGateway $blogGateway;
-    private BlogPermissions $blogPermissions;
-    private Session $session;
-
     public function __construct(
-        BlogGateway $blogGateway,
-        BlogPermissions $blogPermissions,
-        Session $session
+        private readonly BlogGateway $blogGateway,
+        private readonly BlogPermissions $blogPermissions,
+        private readonly Session $session,
+        private readonly Sanitizer $sanitizer,
     ) {
-        $this->blogGateway = $blogGateway;
-        $this->blogPermissions = $blogPermissions;
-        $this->session = $session;
+    }
+
+    /**
+     * Returns a specific blog post.
+     *
+     * @OA\Parameter(name="blogPostId", in="path", @OA\Schema(type="integer"), description="which post to return")
+     * @OA\Response(response="200", description="Success.", @Model(type=BlogPost::class))
+     * @OA\Response(response="404", description="Blog post not found.")
+     * @OA\Tag(name="blog")
+     * @Rest\Get("blog/{blogPostId}", requirements={"blogPostId" = "\d+"})
+     */
+    public function getBlogpostAction(int $blogPostId): Response
+    {
+        $blogPost = $this->blogGateway->getPost($blogPostId);
+
+        if (is_null($blogPost)) {
+            throw new NotFoundHttpException('Blog post not found');
+        }
+
+        $blogPost->content = $this->sanitizer->purifyHtml($blogPost->content);
+
+        return $this->handleView($this->view($blogPost, 200));
     }
 
     /**